Fixed type annotation for workbook.refresh
`workbook.refresh` is implemented to accept both `WorkbookItem`
and `str` as arguments, but the type annotation describes it as
receiving `str`, which can cause false warnings in static analysis.

Since the documentation states that it receives `workbook_item`,
the name of the argument is also changed from `workbook_id` to
`workbook_item`.

Issue: https://github.com/tableau/server-client-python/issues/1318